linting; move ckeditor component
